+++ src/main/java/com/findmymeme/user/repository/UserRepository.java
@@ -1,9 +1,11 @@
-package com.findmymeme.user;
+package com.findmymeme.user.repository;
 
+import com.findmymeme.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
+    boolean existsByUsername(String username);
 }
