+++ src/test/java/com/findmymeme/user/UserRepositoryTest.java
@@ -1,5 +1,7 @@
 package com.findmymeme.user;
 
+import com.findmymeme.user.domain.User;
+import com.findmymeme.user.repository.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
